read_gcs_write_bq_requirements.txt
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-cloud-bigquery==3.10.0
pandas==1.4.2
gcsfs==2023.5.0
pyarrow==12.0.0
db-dtypes==1.1.1
futures==3.1.1